(edit) flatten the 'Dataset' folder and move the components directly under 'components/', and update their imports and references
