--- src/components/Dataset/index.js
@@ -1,9 +1,0 @@
-export { DatasetExplore } from './DatasetExplore'
-export { DatasetPageHeader } from './DatasetPageHeader'
-export { DatasetProcessing } from './DatasetProcessing'
-export { DatasetProcessingError } from './DatasetProcessingError'
-export { DatasetReady } from './DatasetReady'
-export { DatasetRegenerate } from './DatasetRegenerate'
-export { MoveToDatasetButton } from './MoveToDatasetButton'
-export { MoveToDatasetModal } from './MoveToDatasetModal'
-export { ShareDatasetButton } from './ShareDatasetButton'
