+++ src/components/DatasetProcessing.js
@@ -2,9 +2,9 @@ import { Box, Heading, Paragraph } from 'grommet'
 import { useDatasetManager } from 'hooks/useDatasetManager'
 import { useResponsive } from 'hooks/useResponsive'
 import { Column } from 'components/Column'
+import { DatasetExplore } from 'components/DatasetExplore'
 import { IconProcessingGears } from 'components/IconProcessingGears'
 import { Row } from 'components/Row'
-import { DatasetExplore } from './DatasetExplore'
 
 export const DatasetProcessing = ({ dataset }) => {
   const { setResponsive } = useResponsive()
